+++ powerauth-nextstep-model/src/main/java/com/wultra/security/powerauth/lib/nextstep/model/entity/CredentialGenerationParam.java
@@ -45,5 +45,6 @@ public class CredentialGenerationParam {
     private boolean includeSpecialChars;
     @Positive
     private Integer specialCharsCount;
+    private String specialChars;
 
 }